+++ packages/language-support/src/antlr-grammar/CommandLexer.g4
@@ -1,0 +1,9 @@
+lexer grammar CommandLexer;
+
+import CypherLexer;
+
+PARAM : P A R A M S?;
+
+CLEAR: C L E A R;
+
+HISTORY: H I S T O R Y;
