--- packages/language-support/src/antlr-grammar/ConsoleCommandLexer.g4
@@ -1,51 +1,0 @@
-lexer grammar ConsoleCommandLexer;
-
-import CypherLexer;
-
-PARAM : P A R A M S?;
-
-HELP : H E L P; 
-
-PLAY: P L A Y; 
-
-CLEAR: C L E A R;
-
-HISTORY: H I S T O R Y;
-
-AUTO: A U T O ;
-
-CONFIG: C O N F I G;
-
-DBS: D B S;
-
-STYLE: S T Y L E;
-
-GUIDE: G U I D E;
-
-SERVER: S E R V E R;
-
-QUERIES: Q U E R I E S;
-
-SCHEMA: S C H E M A;
-
-SYSINFO: S Y S I N F O;
-
-CONNECT: C O N N E C T;
-
-CHANGEPASSWORD: C H A N G E MINUS P A S S W O R D;
-
-DISCONNECT: D I S C O N N E C T;
-
-STATUS: S T A T U S;
-
-BEGIN: B E G I N;
-
-COMMIT: C O M M I T;
-
-ROLLBACK: R O L L B A C K;
-
-EXIT: E X I T;
-
-SOURCE: S O U R C E;
-
-IMPERSONATE: I M P E R S O N A T E;
